+++ src/main/java/org/angelesyvalientes/api/persistence/repository/TipoIdentificacionRepository.java
@@ -28,7 +28,7 @@ import org.springframework.stereotype.Repository;
  * @see Repository La anotación de Spring que marca esta interfaz como un componente de repositorio.
  */
 @Repository
-public interface TipoIdentificacionRepository  extends JpaRepository<TipoIdentificacion, Long> {
+public interface TipoIdentificacionRepository  extends JpaRepository<TipoIdentificacion, Integer> {
 
     // Aún no se han definido métodos de consulta personalizados aquí.
     // Spring Data JPA proporciona implementaciones para métodos como:
